(BKR-1184) Move TestSuiteResult class into its own file
No functionality change happening here. The class is just moved into its own
file.
